Moves registration form and converter to web/registration folder
