+++ src/main/java/de/hsba/bi/fahrradkurier/web/registration/RegistrationFormConverter.java
@@ -1,16 +1,18 @@
-package de.hsba.bi.fahrradkurier.user;
+package de.hsba.bi.fahrradkurier.web.registration;
 
-import de.hsba.bi.fahrradkurier.Common.AddressEntity;
+import de.hsba.bi.fahrradkurier.common.AddressEntity;
+import de.hsba.bi.fahrradkurier.user.User;
+import de.hsba.bi.fahrradkurier.user.UserRoleEnum;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UserFormConverter {
+public class RegistrationFormConverter {
     private final PasswordEncoder passwordEncoder;
 
-    public User toEntity(UserForm userForm){
+    public User toEntity(RegistrationForm userForm){
         AddressEntity mappedAddressEntity = AddressEntity.builder()
                 .street(userForm.getStreet())
                 .city(userForm.getCity())
